Move #generate_*, #offset_keys, #*crypt methods into EngineModule
